requirements.txt
Flask>=2.3,<3.1
Flask-SQLAlchemy>=3.1,<3.2
Flask-Login>=0.6,<0.7
Flask-WTF>=1.2,<1.3
SQLAlchemy>=2.0,<2.1
python-dotenv>=1.0,<1.1
openai>=1.0,<2.0
werkzeug>=3.0,<3.1
email-validator>=2.0,<2.2
chromadb>=0.4.24,<0.5.0
sentence-transformers>=2.6.0,<3.0
gunicorn>=21.0,<22.0
psycopg2-binary